Finished installer.py
VMwareRemote will now appear in the Start Menu in Windows and in the
Linux app list when installed using the installer.py.
Also the installer now runs in 2 separate subprocesses so the second
part can use the modules installed in the first part.
Also fixed some small bugs in the code (and used os.path.join instead of
pathSeparators for paths...)